Overhall of bound thermal treatment
Nuclides are now defined by the tuple (Z,A,Sab), where
Sab is a string that describes the thermal treatment.
Now, H-1 in H2O, H-1 in ZrH, and free-gas H-1 are separate
nuclides, where previously they were the same. This solves
a problem where two or more nuclides that may be present in
the problem but with different bound thermal treatments.
Bound thermal cross sections are evaluated on a temperature
grid, and different Sab's have different grids. This led to
situations where barnfire had to be run multiple times per
problem. This is completely remedied.

This commit also alleviates issues when more than ten temperatures
are used for a (Z,A) but over different Sab's. There is still
the restriction that a (Z,A,Sab) be associated with at most
ten temperatures. This is a fundamental limitation of the way
MCNP uses material definition strings (ZZAAA.XYz). I have chosen
to associate X with Sab and Y with temperature. Since Y must be
a digit, it can only have ten values, and hence at most ten
temperatures can be used per nuclide. z is a character that is
'c' for non-thermal data and 't' for bound thermal data. Notably,
for ACE files, the bound thermal cross section and its
corresponding non-thermal cross section have the same XY (e.g.,
H-1 in H2O and "lwtr" both have an X of 0, whereas free H-1 has
an X of 9).

Additional changes include improved commenting and referencing of
manuals, especially in materials_util.py.

The function to print out each material in a way conducive to
copying and pasting into an MCNP input deck is also improved,
and now uses consistent XY numbering as the ACE files.

Finally, an option has been added to automatically download
thermal ENDF/B-VII.1 files from BNL.