Changing an elif to an if for printing fission
Some nuclides, such as U-234, will print (6,19), but not (6,18)
by default. Since Readgroupr requires (6,18) to compute the fission
spectrum, if the nuclide is fissionable, (6,18) should always be
in the output GENDF file.

The previous behavior was to add in (6,18) to the GROUPR input
if the nuclide was fissionable but the transfer matrices were not
desired. The new behavior does not depend on whether other transfer
matrices are being used.